testmind/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pymongo==4.10.1
pandas==2.2.3
numpy==2.2.1
scipy==1.15.0
anthropic==0.42.0
openai==1.58.1
python-dotenv==1.0.1
